requirements.txt
httpx==0.27.2
python-dotenv==1.0.1
pandas==2.2.3
plotly==5.24.1
jupyter==1.1.1
ipykernel==6.29.5
